goircam: implement a better anti-aliasing
Instead of drawing boxes, we now compute an average value between a pixel
and its adjacent ones. However the drawPixel() function is extremely slow
because there's no frame buffer and each pixel induces an SPI transaction.

Instead we're filling a full line in memory and rendering it at once in a
single transaction. This remains very fast and much smoother.