goircam: speed up the data transfer
The datasheet clearly indicates that the two subpages are independent,
so it's pointless to try to wait for the right sub-page to come, it
might be different quite from the first downloaded one. Instead now
we just try to download two consecutive distinct pages. This way,
the first image may be read many times if it doesn't change, but
after the first change we'll get two consecutive sub-pages thus a
complete image. This brings the same (low) quality as before, but
significantly faster, which encourages to keep the button pressed.