goircam: don't redraw the whole screen if button is not released
At 8 Hz it's visible that the screen clearing and redrawing starts to
take some time. When the button is held, we know it was already drawn
once, so we don't need to redraw the legends. As such we don't need to
erase it either. By avoiding both the screen clearing and the redrawing
we can double the frame rate, and make it look smoother by not flashing.